+++ app/sprinkles/account/locale/ru_RU/validate.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ *
+ * Russian message token translations for the 'account' sprinkle.
+ *
+ * @package userfrosting\i18n\ru_RU
+ * @author @rendername
+ */
+
+return [
+	"VALIDATE" => [
+        "PASSWORD_MISMATCH" => "Пароли не совпадают.",
+        "USERNAME" => "Имя может состоять только из строчных букв, цифр, '.', '-' и «_»."
+    ]
+];