+++ app/sprinkles/core/locale/ru_RU/validate.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ *
+ * Russian message token translations for the 'core' sprinkle.
+ *
+ * @package userfrosting\i18n\ru_RU
+ * @author @rendername
+ */
+
+return [
+	"VALIDATE" => [
+        "ARRAY" => "Значения для <strong>{{label}}</strong> должны быть элементами массива.",
+        "BOOLEAN" => "Значение <strong>{{label}}</strong> должно быть '0' или '1'.",
+        "INTEGER" => "Значение <strong>{{label}}</strong> должно быть целым.",
+        "INVALID_EMAIL" => "Неправильный email.",
+        "LENGTH_RANGE" => "{{label}} должно быть между {{min}} и {{max}} символов в длину.",
+        "MAX_LENGTH" => "{{label}} должны быть максимально {{max}} символов в длину.",
+        "MIN_LENGTH" => "{{label}} должно быть минимально {{min}} символов в длину.",
+        "NO_LEAD_WS" => "Значение <strong>{{label}}</strong> не может начинаться с пробелов, табуляции или других пробелов.",
+        "NO_TRAIL_WS" => "Значение <strong>{{label}}</strong> не может заканчиваться пробелами, табуляции или другими пробелами.",
+        "RANGE" => "Значение <strong>{{label}}</strong> должно быть между {{min}} и {{max}} симв.",
+        "REQUIRED" => "Пожалуйста, укажите значение для <strong>{{label}}</strong>.",
+        "SPRUNJE" => [
+            "BAD_FILTER" => "<strong>{{name}}</strong> не является допустимым фильтром.",
+            "BAD_LIST" => "<strong>{{name}}</strong> не является допустимым списком.",
+            "BAD_SORT" => "<strong>{{name}}</strong> не является допустимым для сортировки полей."
+        ]
+    ]
+];