+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 - Add chevron to user menu
 - Change "remember me" text
 - Improve table tool buttons
-- Twig extensions now implement `Twig_Extension_GlobalsInterface` as required by https://twig.symfony.com/doc/2.x/advanced.html#id1
+- Twig extensions now implement `Twig_Extension_GlobalsInterface` as required by https://twig.symfony.com/doc/2.x/advanced.html#id1 (#788)
 - Display element based on permissions for group list/info pages 
 - Factor the admin user creation out of migrations and into its own Bakery command (See #778)
 - Bakery `clear-cache` command now clears Twig and router cache (Fix #750)
+- Add Russian translations
 
 ## v4.1.11-alpha
 - Updated `composer/installers` dependency.